Major correction on MSHR allocation and processor stall mechanism
- MSHR allocation:
    - Originally : an MSHR entry is allocated once memory request is
issued.
    - Correction : an MSHR entry should be allocated only if L2 access
is triggered. A L1 read/write hit doesn't need MSHR. There are three
scenarios:
        - L1 read miss
        - L1 write miss
        - L1 write hit, but the requester doesn't have Exclusive state.
Need to access the directory node and trigger invalidation and data
packet.

- Processor stall:
    - Originally: a processor stalls if 1) instruction window (which is
also ROB) is full, or 2) the next is a memory instruction, but no MSHR
is available.
    - Correction: 1) is retained. 2) -> the next is a memory
instruction which requires L2 access. L1 should not be blocked as it
doesn't require a MSHR entry.

- Added: When Request is serviced, before marking an instruction is
ready to commit, we need to compare the instruction type and the block
address in the ROB with the serviced request. The following two cases
should be opted out.

          1) if the serviced request is a READ, but an ROB entry is
Write, although the block address is the same.
          2) Vice Versa

Unimplemented functions in the simulator:

All the following factors may limit the traffic in the network, so as
to interference.

1) Load and store instructions are executed in-order. A single FIFO is
used to model the instruction queue, ROB, and Load/Store Queue (LSQ)
altogether. The disambiguation for a normal LSQ is not modeled. So less
instruction level parallelism is observed.
2) I assume the execution is not bounded by the size of LSQ, given that
the instruction queue is not full.
3) Branch predictor is not modeled.