Add naive multicast with synthetic traffic
1. Correct bufferless router:
   During output assignment, check if (input [i] == null) inside the
loop. Otherwise, flits will be lost

2. Implement multicast synthetic traffic generator. A multicast request
will be split to multiple unicast packet and injected in sequentially.

3. Do not generate memory controller when running synthetic traffic.

Known Issue:
a) The bufferless router latency model.

Currently, the pipeline latency is modeled by link using a fifo in a
non-blocking manner. So, the newly injected flit must experience all
pipeline latency at the source router.

b) Current simulator doesn’t model the synchronization bahavior.

Lock is constructed here. However, this is a very naive implementation.
A simply queue (implemented as a dictionary) is used. No timing model,
synchronization is not associated with cache coherence protocol.
Another word, the synchronization primitive activity is not modeled.

MUST implement this in the future to accurately model the
synchronization behavior for the multithreaded applications.